+++ src/tika_client/_constants.py
@@ -1,0 +1,4 @@
+from typing import Final
+
+# Only compress content which is larger than this
+MIN_COMPRESS_LEN: Final[int] = 1024