+++ src/tika_client/_types.py
@@ -1,0 +1,6 @@
+from typing import Optional
+from typing import Union
+
+MimeType = Optional[str]
+
+RequestContent = Union[str, bytes]